refactor: scene module
