+++ reverie-engine/src/scene/entity.rs
@@ -1,0 +1,3 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
+/// [`Scene`]に登録したエンティティのインデックス
+pub struct EntityIndex(pub hecs::Entity);
